fixup medium redirects and site_hostname env var
